requirements.txt
streamlit
pydub
openai-whisper
torch
python-docx
pytest
